rename db module to models to better reflect its purpose
