+++ src/lib.rs
@@ -1,4 +1,4 @@
-// Need this to expose DB module for tests
+// Need this to expose models module for tests
 pub mod attendance;
-pub mod db;
 pub mod leaderboard;
+pub mod models;
