+++ src/models/member.rs
@@ -1,9 +1,7 @@
 use async_graphql::SimpleObject;
 use sqlx::FromRow;
 
-//Struct for the Member table
 #[derive(FromRow, SimpleObject)]
-
 pub struct Member {
     pub id: i32,
     pub rollno: String,
